feat(frontend): Phase 4 - Ops pages modular refactoring + build_full_html V2 support
Migrated 2 ops pages to V2 directory structure:
- ops-dashboard.js (594→6 files): register + page + ResourceCards + TrendChart + McpHealth + CronStatus
- ops-alerts.js (579→5 files): register + page + RulesTab + HistoryTab + constants

Key improvements:
- ErrorHandler.wrap() on all register.js
- Lazy loading via dynamic import()
- data-action event delegation replacing inline onclick
- TrendChart.buildHorizontalBar injected to avoid circular deps

Critical: Updated app.py build_full_html() to support V2 directory scanning:
- Scans pages/ subdirectories for register.js
- Includes core/ layer scripts in correct dependency order
- Falls back to scanning all .js in directory if no register.js
